refactor: move StandardQR related to Internals.StandardQR
