+++ src/SkiaSharp.QrCode/Internals/StandardQr/ModulePlacer.cs
@@ -2,7 +2,7 @@ using System.Buffers.Binary;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace SkiaSharp.QrCode.Internals;
+namespace SkiaSharp.QrCode.Internals.StandardQr;
 
 /// <summary>
 /// Static utility class for placing QR code modules (patterns and data).
