+++ tests/SkiaSharp.QrCode.Tests/FormatInformationDecoderUnitTest.cs
@@ -1,3 +1,4 @@
+using SkiaSharp.QrCode.Internals.StandardQr;
 using SkiaSharp.QrCode.Internals;
 using SkiaSharp.QrCode.Internals.BinaryDecoders;
