+++ tests/SkiaSharp.QrCode.Tests/OtsuThresholdParityTest.cs
@@ -1,3 +1,4 @@
+using SkiaSharp.QrCode.Internals.StandardQr;
 using SkiaSharp.QrCode.Internals.ImageDecoders;
 
 namespace SkiaSharp.QrCode.Tests;
