+++ tests/SkiaSharp.QrCode.Tests/QRBinaryDecoderUnitTest.cs
@@ -1,3 +1,4 @@
+using SkiaSharp.QrCode.Internals.StandardQr;
 using SkiaSharp.QrCode.Internals.BinaryDecoders;
 using SkiaSharp.QrCode.Internals.BinaryEncoders;
