+++ tests/SkiaSharp.QrCode.Tests/QRBinaryEncoderUnitTest.cs
@@ -1,3 +1,4 @@
+using SkiaSharp.QrCode.Internals.StandardQr;
 using SkiaSharp.QrCode.Internals;
 using SkiaSharp.QrCode.Internals.BinaryEncoders;
 using System.Text;
