+++ tests/SkiaSharp.QrCode.Tests/QRCodeGeneratorUnitTest.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using SkiaSharp.QrCode.Internals.StandardQr;
 using SkiaSharp.QrCode.Internals;
 using static SkiaSharp.QrCode.QRCodeGenerator;
